+++ gazebo/physics/Link.cc
@@ -631,6 +631,12 @@ math::Vector3 Link::GetWorldAngularMomentum() const
 }
 
 //////////////////////////////////////////////////
+math::Vector3 Link::GetWorldAngularMomentum() const
+{
+  return this->GetWorldInertiaMatrix() * this->GetWorldAngularVel();
+}
+
+//////////////////////////////////////////////////
 math::Vector3 Link::GetRelativeForce() const
 {
   return this->GetWorldPose().rot.RotateVectorReverse(this->GetWorldForce());